chore: 패키지 구조 변경
- domain 패키지 제거
- 루트 디렉토리 변경
- exception 패키지 변경

Resolves: #3
