+++ paytogether/src/main/java/com/paytogether/common/api/response/ErrorResponse.java
@@ -1,0 +1,15 @@
+package com.paytogether.common.api.response;
+
+import lombok.Getter;
+
+@Getter
+public class ErrorResponse {
+
+  private final int errorCode;
+  private final String errorMessage;
+
+  public ErrorResponse(int errorCode, String errorMessage) {
+    this.errorMessage = errorMessage;
+    this.errorCode = errorCode;
+  }
+}
