--- paytogether/src/main/java/com/paytogether/common/exception/PayTogetherException.java
@@ -1,13 +1,0 @@
-package com.paytogether.common.exception;
-
-import lombok.Getter;
-
-@Getter
-public class PayTogetherException extends RuntimeException {
-
-  private final PayTogetherErrors error;
-
-  public PayTogetherException(PayTogetherErrors error) {
-    this.error = error;
-  }
-}
