--- paytogether/src/main/java/com/paytogether/domain/member/entity/Gender.java
@@ -1,14 +1,0 @@
-package com.paytogether.domain.member.entity;
-
-import lombok.RequiredArgsConstructor;
-
-
-@RequiredArgsConstructor
-public enum Gender {
-
-  MALE("남"),
-  FEMALE("여"),
-  OTHER("그외");
-
-  private final String explain;
-}
