--- paytogether/src/main/java/com/paytogether/domain/member/repository/MemberRepository.java
@@ -1,10 +1,0 @@
-package com.paytogether.domain.member.repository;
-
-import com.paytogether.domain.member.entity.Member;
-import java.util.Optional;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MemberRepository extends JpaRepository<Member, Long> {
-
-  Optional<Member> findByEmail(String email);
-}
