--- paytogether/src/main/java/com/paytogether/domain/member/service/LoginService.java
@@ -1,10 +1,0 @@
-package com.paytogether.domain.member.service;
-
-public interface LoginService {
-
-  LoginResponse login(String email, String password);
-
-  void logout();
-
-  boolean sessionExists();
-}
