--- paytogether/src/main/java/com/paytogether/domain/member/service/MemberJoinResponse.java
@@ -1,20 +1,0 @@
-package com.paytogether.domain.member.service;
-
-import com.paytogether.domain.member.entity.Member;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class MemberJoinResponse {
-
-  private String email;
-
-  private MemberJoinResponse(String email) {
-    this.email = email;
-  }
-
-  public static MemberJoinResponse from(Member member) {
-    return new MemberJoinResponse(member.getEmail());
-  }
-}
