+++ paytogether/src/main/java/com/paytogether/exception/PayTogetherException.java
@@ -1,0 +1,13 @@
+package com.paytogether.exception;
+
+import lombok.Getter;
+
+@Getter
+public class PayTogetherException extends RuntimeException {
+
+  private final PayTogetherErrors error;
+
+  public PayTogetherException(PayTogetherErrors error) {
+    this.error = error;
+  }
+}
