+++ paytogether/src/main/java/com/paytogether/member/entity/Member.java
@@ -1,0 +1,66 @@
+package com.paytogether.member.entity;
+
+import com.paytogether.common.BaseEntity;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Entity
+public class Member extends BaseEntity {
+
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  @Column(name = "member_id")
+  private Long id;
+
+  private String email;
+
+  private String password;
+
+  private String name;
+
+  private int age;
+
+  @Enumerated(EnumType.STRING)
+  private Gender gender;
+
+  private String address;
+
+  private String phoneNumber;
+
+  @Enumerated(EnumType.STRING)
+  private MemberRole role;
+
+  private boolean isActive = true;
+
+  @Builder
+  private Member(String email, String password, String name, int age, Gender gender, String address,
+      String phoneNumber, MemberRole role) {
+    this.email = email;
+    this.password = password;
+    this.name = name;
+    this.age = age;
+    this.gender = gender;
+    this.address = address;
+    this.phoneNumber = phoneNumber;
+    this.role = role;
+  }
+
+  public static Member createMember(String email, String encodedPassword, String name, int age,
+      Gender gender, String address, String phoneNumber) {
+    return Member.builder()
+        .email(email).password(encodedPassword).name(name).age(age)
+        .gender(gender).address(address).phoneNumber(phoneNumber).role(MemberRole.MEMBER)
+        .build();
+  }
+}
