+++ paytogether/src/main/java/com/paytogether/member/entity/MemberRole.java
@@ -1,0 +1,13 @@
+package com.paytogether.member.entity;
+
+import lombok.RequiredArgsConstructor;
+
+
+@RequiredArgsConstructor
+public enum MemberRole {
+
+  MEMBER("회원"),
+  ADMIN("관리자");
+
+  private final String explain;
+}
