+++ paytogether/src/main/java/com/paytogether/member/repository/MemberRepository.java
@@ -1,0 +1,10 @@
+package com.paytogether.member.repository;
+
+import com.paytogether.member.entity.Member;
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+
+  Optional<Member> findByEmail(String email);
+}
