+++ paytogether/src/main/java/com/paytogether/member/service/LoginService.java
@@ -1,0 +1,10 @@
+package com.paytogether.member.service;
+
+public interface LoginService {
+
+  LoginResponse login(String email, String password);
+
+  void logout();
+
+  boolean sessionExists();
+}
