--- paytogether/src/test/java/com/paytogether/domain/member/service/TestMemberConstant.java
@@ -1,6 +1,0 @@
-package com.paytogether.domain.member.service;
-
-public class TestMemberConstant {
-
-  protected static final String LOGIN_MEMBER = "LOGIN_MEMBER";
-}
