+++ paytogether/src/test/java/com/paytogether/member/service/TestMemberConstant.java
@@ -1,0 +1,6 @@
+package com.paytogether.member.service;
+
+public class TestMemberConstant {
+
+  protected static final String LOGIN_MEMBER = "LOGIN_MEMBER";
+}
